refactor: remove shallow re-export modules and a dead branch
