+++ src/utils/device.ts
@@ -39,14 +39,6 @@ type DeviceSelectionContext = {
   simulatorSetPath?: string;
 };
 
-export function normalizePlatformSelector(
-  platform: PlatformSelector | undefined,
-): PlatformSelector | undefined {
-  // Single normalization hook for platform selectors. Current CLI parsing already
-  // yields canonical values, but Apple-family routing still depends on one shared point.
-  return platform;
-}
-
 export function isApplePlatform(
   platform: Platform | PlatformSelector | undefined,
 ): platform is ApplePlatform | 'apple' {
